Condense AddCardCommand, AddChapterCommand and AddModuleCommand into single AddCommand
