+++ src/main/java/ui/Ui.java
@@ -2,14 +2,13 @@ package ui;
 
 import access.Access;
 
-import commands.AddCardCommand;
-import commands.HelpCommand;
-import commands.ReviseCommand;
-import commands.ListCommand;
-import commands.ExitCommand;
+import commands.AddCommand;
 import commands.EditCommand;
+import commands.ExitCommand;
+import commands.HelpCommand;
+import commands.ListCommand;
 import commands.RemoveCommand;
-
+import commands.ReviseCommand;
 import manager.card.Card;
 import manager.chapter.Chapter;
 import manager.module.Module;
@@ -110,7 +109,7 @@ public class Ui {
         out.println("1. " + ListCommand.MESSAGE_USAGE);
         out.println("2. " + ReviseCommand.MESSAGE_USAGE);
         out.println("3. " + HelpCommand.MESSAGE_USAGE);
-        out.println("4. " + AddCardCommand.MESSAGE_USAGE);
+        out.println("4. " + AddCommand.MESSAGE_USAGE);
         out.println("5. " + ExitCommand.MESSAGE_USAGE);
         out.println("6. " + EditCommand.MESSAGE_USAGE);
         out.println("7. " + RemoveCommand.MESSAGE_USAGE);
